added vozila update delete and insert

+++ PPPK/PPPK_Web/PPPK_Web/CustomValidators/KilometriValidator.cs
@@ -10,6 +10,6 @@ namespace PPPK_Web.CustomValidators
     public class KilometriValidator : ValidationAttribute
     {
         public override bool IsValid(object value)
-            => (value != null && Validators.validKilometar(Convert.ToDouble(value))) ? true : false;
+            => (value != null && Validators.validKilometar(Convert.ToDecimal(value))) ? true : false;
     }
 }
